backend/requirements.txt
Flask
flask-mysqldb
flask-jwt-extended
python-dotenv
mysql-connector-python